requirements.txt
influxdb-client
pandas
opencv-python
spidev
timeout-decorator
pyserial
cobs